// src/shared/api/generated/model/depth2.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * 아이랑 - 홈 화면 API
 * OpenAPI spec version: 0.0.1
 */

export type Depth2 = typeof Depth2[keyof typeof Depth2];


export const Depth2 = {
  HS01: 'HS01',
  HS02: 'HS02',
  HS03: 'HS03',
  HS04: 'HS04',
  NA01: 'NA01',
  NA02: 'NA02',
  NA03: 'NA03',
  NA04: 'NA04',
  NA05: 'NA05',
  EX01: 'EX01',
  EX02: 'EX02',
  EX03: 'EX03',
  EX04: 'EX04',
  EX05: 'EX05',
  EX06: 'EX06',
  EX07: 'EX07',
} as const;